BugFix: LDA (zp,X)

+++ src/dev/mem/mmio.c
@@ -1401,13 +1401,13 @@ INLINE uint8_t _src_ind_dis(void) {
  effective address is word in (LL + X, LL + X + 1), inc. without carry: C.w($00LL + X)
  **/
 INLINE uint16_t _addr_ind_X(void) {
-    return memread16( _fetch() + m6502.X );
+    return memread16( (uint8_t)(_fetch() + m6502.X) );
 }
 INLINE uint16_t _addr_ind_X_rd_dbg(void) {
-    return _memread16_dbg( _fetch() + m6502.X );
+    return _memread16_dbg( (uint8_t)(_fetch() + m6502.X) );
 }
 INLINE uint16_t _addr_ind_X_dbg(void) {
-    uint16_t addr = _memread16_dbg(_fetch() + m6502.X);
+    uint16_t addr = _memread16_dbg( (uint8_t)(_fetch() + m6502.X));
     check_mem_wr_bp(addr); // write debug on the target address
     return addr;
 }
@@ -1415,7 +1415,7 @@ INLINE uint16_t _addr_ind_X_dis(void) {
     _disPrintf(disassembly.oper, sizeof(disassembly.oper), "($%02X,X)", memread8(m6502.PC) );
     _disPrintf(disassembly.comment, sizeof(disassembly.comment), "ind_addr:%04X", memread16( memread8(m6502.PC) + m6502.X) );
 
-    return memread16( _fetch_dis() + m6502.X );
+    return memread16( (uint8_t)(_fetch_dis() + m6502.X) );
 }
 INLINE uint8_t _src_X_ind(void) {
     return _memread( _addr_ind_X() );
